V1.1.0 - refactor config to own module
