+++ src/cli.rs
@@ -1,14 +1,9 @@
-
 use super::{
     Color,
-    constants::{APPLICATION, ORGANIZATION, QUALIFIER}
+    constants::{APPLICATION, ORGANIZATION, QUALIFIER},
 };
 use clap::{Parser, Subcommand};
 use clap_complete::Shell;
-use directories::ProjectDirs;
-use serde::{Deserialize, Serialize};
-use serde_json;
-use std::fs;
 
 #[derive(Parser)]
 #[command(author, version, about)]
@@ -88,45 +83,4 @@ pub enum SolidMode {
 
 pub enum Mode {
     Solid(SolidMode),
-}
-
-pub struct Config {
-    // From CLI options
-    verbose: bool,
-
-    // User configurations
-    frequency: Option<String>,
-    generators: Option<Vec<String>>,
-}
-
-#[derive(Default, Deserialize)]
-struct UserConfig {
-    frequency: Option<String>,
-    generators: Option<Vec<String>>,
-}
-
-impl Config {
-    pub fn new(verbose: bool) -> Self {
-        let UserConfig { frequency, generators } = Config::read_config_file_if_exists();
-        Self { frequency, generators, verbose }
-    }
-
-    pub fn print_if_verbose(&self, message: &str) {
-        if self.verbose {
-            println!("{}", message);
-        }
-    }
-
-    fn read_config_file_if_exists() -> UserConfig {
-        let data = ProjectDirs::from(QUALIFIER, ORGANIZATION, APPLICATION)
-            .map(|dirs| dirs.data_dir().join("config.json"))
-            .filter(|path| path.exists())
-            .and_then(|path| fs::read_to_string(path).ok());
-        if let Some(data) = data {
-            // TODO: add proper error handling and throwing here
-            serde_json::from_str(&data).expect("Config should be correct, TODO: add proper error here")
-        } else {
-            UserConfig::default()
-        }
-    }
 }
